Fix errors after switching to rust 1.0 alpha

+++ src/http/mod.rs
@@ -6,8 +6,8 @@ mod client;
 
 #[deriving(Show, PartialEq, Copy)]
 pub struct ResponseHeader {
-    status: int,
-    content_length: int,
+    status: i32,
+    content_length: isize,
 }
 
 #[deriving(Show, PartialEq)]
@@ -15,6 +15,6 @@ pub struct RequestHeader {
     method: String,
     request_uri: String,
     http_version: String,
-    content_length: int,
+    content_length: i32,
 }
